feat: UI 디자인 수정
