+++ app/src/main/java/net/developia/livartc/mypage/membership/MemberShipFragment.kt
@@ -14,7 +14,10 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import net.developia.livartc.R
+import net.developia.livartc.databinding.FragmentAccountBinding
+import net.developia.livartc.databinding.FragmentMembershipBinding
 import net.developia.livartc.login.TokenManager
+import net.developia.livartc.main.MyPurchaseFragment
 import net.developia.livartc.mypage.PopUp
 import net.developia.livartc.mypage.dto.MemberGradeDto
 import net.developia.livartc.mypage.dto.PopUpDto
@@ -32,6 +35,7 @@ import kotlin.coroutines.suspendCoroutine
 class MemberShipFragment : Fragment() {
     private lateinit var linearLayout: LinearLayout
     private lateinit var membershipCardImageView: ImageView
+    lateinit var binding: FragmentMembershipBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -45,6 +49,42 @@ class MemberShipFragment : Fragment() {
             val result = fetchDataFromServer()
             updateUI(result)
         }
+
+
+        binding = FragmentMembershipBinding.inflate(inflater, container, false)
+
+        // 1. my_info 버튼 클릭시 새로운 액티비티로 이동
+        val myInfoLayout = view.findViewById<LinearLayout>(R.id.myinfo_tab)
+        myInfoLayout.setOnClickListener {
+            startActivity(Intent(requireContext(), MyInfoActivity::class.java))
+        }
+
+        // 2. my_review 버튼 클릭시 새로운 액티비티로 이동
+        val myReviewLayout = view.findViewById<LinearLayout>(R.id.myreview_tab)
+        myReviewLayout.setOnClickListener {
+            startActivity(Intent(requireContext(), MyReviewActivity::class.java))
+        }
+
+        // 3. my_review 버튼 클릭시 새로운 액티비티로 이동
+        val myOrderLayout = view.findViewById<LinearLayout>(R.id.myorder_tab)
+        myOrderLayout.setOnClickListener {
+            startActivity(Intent(requireContext(), MyOrderActivity::class.java))
+        }
+
+        // 4. logout 버튼 클릭시 => 로그아웃하고 처음 화면으로 이동하기
+        // 일단 팝업 뜨는 버튼으로!!
+        val logoutLayout = view.findViewById<LinearLayout>(R.id.logout_tab)
+        logoutLayout.setOnClickListener {
+            // startActivity(Intent(requireContext(), LogoutActivity::class.java))
+            // 팝업 뜨는 예시
+            val tmp = PopUpDto (
+                "Commenter",
+                "https://github.com/livarter/android/assets/77563814/a346484d-31df-477e-a43c-690aec02c63f",
+                "축하합니다! 뱃지를 발급하였습니다!"
+            )
+            PopUp().show(tmp, requireContext())
+        }
+
         return view
     }
 
@@ -80,5 +120,18 @@ class MemberShipFragment : Fragment() {
         Glide.with(requireContext())
             .load(result.image)
             .into(membershipCardImageView)
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.myorderTab.setOnClickListener {
+            parentFragmentManager.beginTransaction().apply{
+                replace(R.id.main_container, MyPurchaseFragment())
+                addToBackStack(null)
+                commit()
+            }
+        }
     }
 }
