+++ app/src/main/res/layout/fragment_membership.xml
@@ -7,16 +7,172 @@
     android:background="@color/cardview_light_background"
     tools:context=".mypage.membership.MemberShipFragment" >
 
-    <LinearLayout
-        android:id="@+id/profile_membership"
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-        <ImageView
-            android:id="@+id/membership_card"
+        android:layout_height="match_parent">
+
+        <LinearLayout
+            android:id="@+id/profile_membership"
             android:layout_width="match_parent"
-            android:layout_height="200dp"
-            android:layout_marginBottom="30dp"
-            android:src="@drawable/_02_home_brand_kind" />
-    </LinearLayout>
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+        <ImageView
+                android:id="@+id/membership_card"
+                android:layout_width="match_parent"
+                android:layout_height="200dp"
+                android:layout_marginBottom="30dp"
+                android:layout_marginTop="10dp"
+                android:src="@drawable/_02_home_brand_kind" />
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical">
+
+                <!-- 나의 정보 -->
+                <LinearLayout
+                    android:id="@+id/myinfo_tab"
+                    android:layout_width="match_parent"
+                    android:layout_height="60dp"
+                    android:layout_margin="10dp"
+                    android:orientation="horizontal">
+
+                    <ImageView
+                        android:id="@+id/myinfo_icon"
+                        android:layout_width="50dp"
+                        android:layout_height="match_parent"
+                        android:layout_margin="15dp"
+                        tools:src="@drawable/ic_mypage" />
+
+                    <LinearLayout
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_margin="10dp"
+                        android:orientation="vertical">
+
+                        <TextView
+                            android:id="@+id/myinfo_title"
+                            android:layout_width="105dp"
+                            android:layout_height="match_parent"
+                            android:text="나의 정보" />
+                        <TextView
+                            android:id="@+id/myinfo_subtitle"
+                            android:layout_width="match_parent"
+                            android:layout_height="match_parent"
+                            android:text="나의 정보를 조회하고 수정합니다" />
+                    </LinearLayout>
+                </LinearLayout>
+
+                <!-- 나의 리뷰 -->
+                <LinearLayout
+                    android:id="@+id/myreview_tab"
+                    android:layout_width="match_parent"
+                    android:layout_height="60dp"
+                    android:layout_margin="10dp"
+                    android:orientation="horizontal">
+
+                    <ImageView
+                        android:id="@+id/myreview_icon"
+                        android:layout_width="50dp"
+                        android:layout_height="match_parent"
+                        android:layout_margin="15dp"
+                        tools:src="@drawable/ic_write" />
+                    <LinearLayout
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_margin="10dp"
+                        android:orientation="vertical">
+                        <TextView
+                            android:id="@+id/myreview_title"
+                            android:layout_width="105dp"
+                            android:layout_height="match_parent"
+                            android:text="나의 리뷰" />
+
+                        <TextView
+                            android:id="@+id/myreview_subtitle"
+                            android:layout_width="match_parent"
+                            android:layout_height="match_parent"
+                            android:text="나의 리뷰를 조회하고 수정합니다" />
+                    </LinearLayout>
+                </LinearLayout>
+
+                <!-- 나의 주문 -->
+                <LinearLayout
+                    android:id="@+id/myorder_tab"
+                    android:layout_width="match_parent"
+                    android:layout_height="60dp"
+                    android:layout_margin="10dp"
+                    android:orientation="horizontal">
+
+                    <ImageView
+                        android:id="@+id/myorder_icon"
+                        android:layout_width="50dp"
+                        android:layout_height="match_parent"
+                        android:layout_margin="15dp"
+                        tools:src="@drawable/ic_cart" />
+
+                    <LinearLayout
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_margin="10dp"
+                        android:orientation="vertical">
+
+                        <TextView
+                            android:id="@+id/myorder_title"
+                            android:layout_width="105dp"
+                            android:layout_height="match_parent"
+                            android:text="나의 주문" />
+
+                        <TextView
+                            android:id="@+id/myorder_subtitle"
+                            android:layout_width="match_parent"
+                            android:layout_height="match_parent"
+                            android:text="나의 정보를 조회하고 수정합니다" />
+                    </LinearLayout>
+                </LinearLayout>
+
+                <LinearLayout
+                    android:id="@+id/logout_tab"
+                    android:layout_width="match_parent"
+                    android:layout_height="60dp"
+                    android:layout_margin="10dp"
+                    android:orientation="horizontal">
+
+                    <ImageView
+                        android:id="@+id/logout_icon"
+                        android:layout_width="50dp"
+                        android:layout_height="match_parent"
+                        android:layout_margin="15dp"
+                        tools:src="@drawable/ic_logout" />
+
+                    <LinearLayout
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_margin="10dp"
+                        android:orientation="vertical">
+
+                        <TextView
+                            android:id="@+id/logout_title"
+                            android:layout_width="105dp"
+                            android:layout_height="match_parent"
+                            android:text="로그아웃" />
+
+                        <TextView
+                            android:id="@+id/logout_subtitle"
+                            android:layout_width="match_parent"
+                            android:layout_height="match_parent"
+                            android:text="로그아웃합니다" />
+                    </LinearLayout>
+                </LinearLayout>
+            </LinearLayout>
+
+        </LinearLayout>
+
+
+
+
+    </ScrollView>
+
+
 </FrameLayout>
